+++ packages/proto/src/gen/cosmwasm.wasm.v1.abstract.rs
@@ -36,10 +36,6 @@ impl ::prost::Name for AccessConfig {
 #[derive(Clone, PartialEq, :: prost :: Message)]
 pub struct Params {
     #[prost(message, optional, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub code_upload_access: ::core::option::Option<AccessConfig>,
     #[prost(enumeration = "AccessType", tag = "2")]
     pub instantiate_default_permission: i32,
@@ -64,10 +60,6 @@ pub struct CodeInfo {
     pub creator: ::prost::alloc::string::String,
     #[doc = " InstantiateConfig access control to apply on contract creation, optional"]
     #[prost(message, optional, tag = "5")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub instantiate_config: ::core::option::Option<AccessConfig>,
 }
 impl ::prost::Name for CodeInfo {
@@ -81,17 +73,8 @@ impl ::prost::Name for CodeInfo {
 #[derive(:: serde :: Serialize, :: serde :: Deserialize)]
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, :: prost :: Message)]
-pub struct ContractInfo<
-    A: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
-> {
+pub struct ContractInfo<A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name>
+{
     #[doc = " CodeID is the reference to the stored Wasm code"]
     #[prost(uint64, tag = "1")]
     pub code_id: u64,
@@ -106,33 +89,16 @@ pub struct ContractInfo<
     pub label: ::prost::alloc::string::String,
     #[doc = " Created Tx position when the contract was instantiated."]
     #[prost(message, optional, tag = "5")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub created: ::core::option::Option<AbsoluteTxPosition>,
     #[prost(string, tag = "6")]
     pub ibc_port_id: ::prost::alloc::string::String,
     #[doc = " Extension is an extension point to store custom metadata within the"]
     #[doc = " persistence model."]
     #[prost(message, optional, tag = "7")]
-    #[serde(
-        serialize_with = "crate::any::option::serialize",
-        deserialize_with = "crate::any::option::deserialize"
-    )]
-    pub extension: ::core::option::Option<A>,
+    pub extension: ::core::option::Option<crate::any::Any<A>>,
 }
-impl<
-        A: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
-    > ::prost::Name for ContractInfo<A>
+impl<A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name> ::prost::Name
+    for ContractInfo<A>
 {
     const NAME: &'static str = "ContractInfo";
     const PACKAGE: &'static str = "cosmwasm.wasm.v1";
@@ -152,10 +118,6 @@ pub struct ContractCodeHistoryEntry {
     pub code_id: u64,
     #[doc = " Updated Tx position when the operation was executed."]
     #[prost(message, optional, tag = "3")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub updated: ::core::option::Option<AbsoluteTxPosition>,
     #[prost(bytes = "vec", tag = "4")]
     pub msg: ::prost::alloc::vec::Vec<u8>,
@@ -319,10 +281,6 @@ impl ContractCodeHistoryOperationType {
 pub struct StoreCodeAuthorization {
     #[doc = " Grants for code upload"]
     #[prost(message, repeated, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub grants: ::prost::alloc::vec::Vec<CodeGrant>,
 }
 impl ::prost::Name for StoreCodeAuthorization {
@@ -338,52 +296,16 @@ impl ::prost::Name for StoreCodeAuthorization {
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, :: prost :: Message)]
 pub struct ContractExecutionAuthorization<
-    A: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
-    BB: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
+    A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
+    BB: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
 > {
     #[doc = " Grants for contract executions"]
     #[prost(message, repeated, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub grants: ::prost::alloc::vec::Vec<ContractGrant<A, BB>>,
 }
 impl<
-        A: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
-        BB: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
+        A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
+        BB: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
     > ::prost::Name for ContractExecutionAuthorization<A, BB>
 {
     const NAME: &'static str = "ContractExecutionAuthorization";
@@ -398,52 +320,16 @@ impl<
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, :: prost :: Message)]
 pub struct ContractMigrationAuthorization<
-    A: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
-    BB: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
+    A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
+    BB: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
 > {
     #[doc = " Grants for contract migrations"]
     #[prost(message, repeated, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub grants: ::prost::alloc::vec::Vec<ContractGrant<A, BB>>,
 }
 impl<
-        A: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
-        BB: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
+        A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
+        BB: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
     > ::prost::Name for ContractMigrationAuthorization<A, BB>
 {
     const NAME: &'static str = "ContractMigrationAuthorization";
@@ -465,10 +351,6 @@ pub struct CodeGrant {
     #[doc = " on contract creation."]
     #[doc = " Optional"]
     #[prost(message, optional, tag = "2")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub instantiate_permission: ::core::option::Option<AccessConfig>,
 }
 impl ::prost::Name for CodeGrant {
@@ -484,24 +366,8 @@ impl ::prost::Name for CodeGrant {
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, :: prost :: Message)]
 pub struct ContractGrant<
-    A: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
-    BB: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
+    A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
+    BB: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
 > {
     #[doc = " Contract is the bech32 address of the smart contract"]
     #[prost(string, tag = "1")]
@@ -509,40 +375,16 @@ pub struct ContractGrant<
     #[doc = " Limit defines execution limits that are enforced and updated when the grant"]
     #[doc = " is applied. When the limit lapsed the grant is removed."]
     #[prost(message, optional, tag = "2")]
-    #[serde(
-        serialize_with = "crate::any::option::serialize",
-        deserialize_with = "crate::any::option::deserialize"
-    )]
-    pub limit: ::core::option::Option<A>,
+    pub limit: ::core::option::Option<crate::any::Any<A>>,
     #[doc = " Filter define more fine-grained control on the message payload passed"]
     #[doc = " to the contract in the operation. When no filter applies on execution, the"]
     #[doc = " operation is prohibited."]
     #[prost(message, optional, tag = "3")]
-    #[serde(
-        serialize_with = "crate::any::option::serialize",
-        deserialize_with = "crate::any::option::deserialize"
-    )]
-    pub filter: ::core::option::Option<BB>,
+    pub filter: ::core::option::Option<crate::any::Any<BB>>,
 }
 impl<
-        A: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
-        BB: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
+        A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
+        BB: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
     > ::prost::Name for ContractGrant<A, BB>
 {
     const NAME: &'static str = "ContractGrant";
@@ -660,53 +502,19 @@ impl ::prost::Name for AcceptedMessagesFilter {
 #[derive(:: serde :: Serialize, :: serde :: Deserialize)]
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, :: prost :: Message)]
-pub struct GenesisState<
-    A: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
-> {
+pub struct GenesisState<A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name>
+{
     #[prost(message, optional, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub params: ::core::option::Option<Params>,
     #[prost(message, repeated, tag = "2")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub codes: ::prost::alloc::vec::Vec<Code>,
     #[prost(message, repeated, tag = "3")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub contracts: ::prost::alloc::vec::Vec<Contract<A>>,
     #[prost(message, repeated, tag = "4")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub sequences: ::prost::alloc::vec::Vec<Sequence>,
 }
-impl<
-        A: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
-    > ::prost::Name for GenesisState<A>
+impl<A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name> ::prost::Name
+    for GenesisState<A>
 {
     const NAME: &'static str = "GenesisState";
     const PACKAGE: &'static str = "cosmwasm.wasm.v1";
@@ -722,10 +530,6 @@ pub struct Code {
     #[prost(uint64, tag = "1")]
     pub code_id: u64,
     #[prost(message, optional, tag = "2")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub code_info: ::core::option::Option<CodeInfo>,
     #[prost(bytes = "vec", tag = "3")]
     pub code_bytes: ::prost::alloc::vec::Vec<u8>,
@@ -744,49 +548,18 @@ impl ::prost::Name for Code {
 #[derive(:: serde :: Serialize, :: serde :: Deserialize)]
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, :: prost :: Message)]
-pub struct Contract<
-    A: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
-> {
+pub struct Contract<A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name> {
     #[prost(string, tag = "1")]
     pub contract_address: ::prost::alloc::string::String,
     #[prost(message, optional, tag = "2")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub contract_info: ::core::option::Option<ContractInfo<A>>,
     #[prost(message, repeated, tag = "3")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub contract_state: ::prost::alloc::vec::Vec<Model>,
     #[prost(message, repeated, tag = "4")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub contract_code_history: ::prost::alloc::vec::Vec<ContractCodeHistoryEntry>,
 }
-impl<
-        A: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
-    > ::prost::Name for Contract<A>
+impl<A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name> ::prost::Name
+    for Contract<A>
 {
     const NAME: &'static str = "Contract";
     const PACKAGE: &'static str = "cosmwasm.wasm.v1";
@@ -892,10 +665,6 @@ pub struct StoreCodeProposal {
     pub wasm_byte_code: ::prost::alloc::vec::Vec<u8>,
     #[doc = " InstantiatePermission to apply on contract creation, optional"]
     #[prost(message, optional, tag = "7")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub instantiate_permission: ::core::option::Option<AccessConfig>,
     #[doc = " UnpinCode code on upload, optional"]
     #[prost(bool, tag = "8")]
@@ -1215,10 +984,6 @@ pub struct AccessConfigUpdate {
     pub code_id: u64,
     #[doc = " InstantiatePermission to apply to the set of code ids"]
     #[prost(message, optional, tag = "2")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub instantiate_permission: ::core::option::Option<AccessConfig>,
 }
 impl ::prost::Name for AccessConfigUpdate {
@@ -1245,10 +1010,6 @@ pub struct UpdateInstantiateConfigProposal {
     #[doc = " AccessConfigUpdate contains the list of code ids and the access config"]
     #[doc = " to be applied."]
     #[prost(message, repeated, tag = "3")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub access_config_updates: ::prost::alloc::vec::Vec<AccessConfigUpdate>,
 }
 impl ::prost::Name for UpdateInstantiateConfigProposal {
@@ -1280,10 +1041,6 @@ pub struct StoreAndInstantiateContractProposal {
     pub wasm_byte_code: ::prost::alloc::vec::Vec<u8>,
     #[doc = " InstantiatePermission to apply on contract creation, optional"]
     #[prost(message, optional, tag = "5")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub instantiate_permission: ::core::option::Option<AccessConfig>,
     #[doc = " UnpinCode code on upload, optional"]
     #[prost(bool, tag = "6")]
@@ -1342,37 +1099,16 @@ impl ::prost::Name for QueryContractInfoRequest {
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, :: prost :: Message)]
 pub struct QueryContractInfoResponse<
-    A: Clone
-        + PartialEq
-        + Default
-        + Send
-        + Sync
-        + prost::Message
-        + serde::Serialize
-        + serde::de::DeserializeOwned
-        + prost::Name,
+    A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name,
 > {
     #[doc = " address is the address of the contract"]
     #[prost(string, tag = "1")]
     pub address: ::prost::alloc::string::String,
     #[prost(message, optional, tag = "2")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub contract_info: ::core::option::Option<ContractInfo<A>>,
 }
-impl<
-        A: Clone
-            + PartialEq
-            + Default
-            + Send
-            + Sync
-            + prost::Message
-            + serde::Serialize
-            + serde::de::DeserializeOwned
-            + prost::Name,
-    > ::prost::Name for QueryContractInfoResponse<A>
+impl<A: Clone + PartialEq + Default + Send + Sync + prost::Message + prost::Name> ::prost::Name
+    for QueryContractInfoResponse<A>
 {
     const NAME: &'static str = "QueryContractInfoResponse";
     const PACKAGE: &'static str = "cosmwasm.wasm.v1";
@@ -1407,10 +1143,6 @@ impl ::prost::Name for QueryContractHistoryRequest {
 #[derive(Clone, PartialEq, :: prost :: Message)]
 pub struct QueryContractHistoryResponse {
     #[prost(message, repeated, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub entries: ::prost::alloc::vec::Vec<ContractCodeHistoryEntry>,
     #[doc = " pagination defines the pagination in the response."]
     #[prost(message, optional, tag = "2")]
@@ -1490,10 +1222,6 @@ impl ::prost::Name for QueryAllContractStateRequest {
 #[derive(Clone, PartialEq, :: prost :: Message)]
 pub struct QueryAllContractStateResponse {
     #[prost(message, repeated, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub models: ::prost::alloc::vec::Vec<Model>,
     #[doc = " pagination defines the pagination in the response."]
     #[prost(message, optional, tag = "2")]
@@ -1608,10 +1336,6 @@ pub struct CodeInfoResponse {
     #[prost(bytes = "vec", tag = "3")]
     pub data_hash: ::prost::alloc::vec::Vec<u8>,
     #[prost(message, optional, tag = "6")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub instantiate_permission: ::core::option::Option<AccessConfig>,
 }
 impl ::prost::Name for CodeInfoResponse {
@@ -1627,10 +1351,6 @@ impl ::prost::Name for CodeInfoResponse {
 #[derive(Clone, PartialEq, :: prost :: Message)]
 pub struct QueryCodeResponse {
     #[prost(message, optional, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub code_info: ::core::option::Option<CodeInfoResponse>,
     #[prost(bytes = "vec", tag = "2")]
     pub data: ::prost::alloc::vec::Vec<u8>,
@@ -1664,10 +1384,6 @@ impl ::prost::Name for QueryCodesRequest {
 #[derive(Clone, PartialEq, :: prost :: Message)]
 pub struct QueryCodesResponse {
     #[prost(message, repeated, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::vec::generic_serialize",
-        deserialize_with = "crate::any::vec::generic_deserialize"
-    )]
     pub code_infos: ::prost::alloc::vec::Vec<CodeInfoResponse>,
     #[doc = " pagination defines the pagination in the response."]
     #[prost(message, optional, tag = "2")]
@@ -1735,10 +1451,6 @@ impl ::prost::Name for QueryParamsRequest {
 pub struct QueryParamsResponse {
     #[doc = " params defines the parameters of the module."]
     #[prost(message, optional, tag = "1")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub params: ::core::option::Option<Params>,
 }
 impl ::prost::Name for QueryParamsResponse {
@@ -1802,10 +1514,6 @@ pub struct MsgStoreCode {
     #[doc = " InstantiatePermission access control to apply on contract creation,"]
     #[doc = " optional"]
     #[prost(message, optional, tag = "5")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub instantiate_permission: ::core::option::Option<AccessConfig>,
 }
 impl ::prost::Name for MsgStoreCode {
@@ -2104,10 +1812,6 @@ pub struct MsgUpdateInstantiateConfig {
     pub code_id: u64,
     #[doc = " NewInstantiatePermission is the new access control"]
     #[prost(message, optional, tag = "3")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub new_instantiate_permission: ::core::option::Option<AccessConfig>,
 }
 impl ::prost::Name for MsgUpdateInstantiateConfig {
@@ -2143,10 +1847,6 @@ pub struct MsgUpdateParams {
     #[doc = ""]
     #[doc = " NOTE: All parameters must be supplied."]
     #[prost(message, optional, tag = "2")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub params: ::core::option::Option<Params>,
 }
 impl ::prost::Name for MsgUpdateParams {
@@ -2302,10 +2002,6 @@ pub struct MsgStoreAndInstantiateContract {
     pub wasm_byte_code: ::prost::alloc::vec::Vec<u8>,
     #[doc = " InstantiatePermission to apply on contract creation, optional"]
     #[prost(message, optional, tag = "4")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub instantiate_permission: ::core::option::Option<AccessConfig>,
     #[doc = " UnpinCode code on upload, optional. As default the uploaded contract is"]
     #[doc = " pinned to cache."]
@@ -2445,10 +2141,6 @@ pub struct MsgStoreAndMigrateContract {
     pub wasm_byte_code: ::prost::alloc::vec::Vec<u8>,
     #[doc = " InstantiatePermission to apply on contract creation, optional"]
     #[prost(message, optional, tag = "3")]
-    #[serde(
-        serialize_with = "crate::any::option::generic_serialize",
-        deserialize_with = "crate::any::option::generic_deserialize"
-    )]
     pub instantiate_permission: ::core::option::Option<AccessConfig>,
     #[doc = " Contract is the address of the smart contract"]
     #[prost(string, tag = "4")]
