+++ proto-build/src/parser/commands/load_and_patch_any.rs
@@ -1,17 +1,12 @@
 use crate::parser::consts::GENERICS;
-use crate::parser::utils::common::{
-    create_punctuated, fields_as_named, gen_generic, item_as_struct,
-};
-use crate::parser::utils::gen_type_param::gen_type_param;
-use crate::parser::utils::is_important::{is_important, FoundEnclosure};
+use crate::parser::utils::common::{fields_as_named, item_as_struct};
+use crate::parser::utils::gen_type_param::{gen_any, gen_type_param};
+use crate::parser::utils::is_important::is_important;
 use glob::glob;
-use proc_macro2::{Ident, Literal, Punct, Spacing, Span, TokenStream};
-use quote::TokenStreamExt;
 use std::collections::BTreeMap;
 use std::fs;
 use std::path::{Path, PathBuf};
-use syn::token::Paren;
-use syn::{AttrStyle, Attribute, File, GenericParam, Item, MacroDelimiter, Meta, MetaList};
+use syn::{File, GenericParam, Item};
 
 pub fn load_and_patch_any(out_dir: &Path) -> BTreeMap<String, (File, BTreeMap<String, usize>)> {
     // Map all file ASTs
@@ -27,11 +22,6 @@ pub fn load_and_patch_any(out_dir: &Path) -> BTreeMap<String, (File, BTreeMap<St
     for src in src_files {
         let current_file = fs::read_to_string(&src).unwrap();
 
-        // // Filter files that don't have `Any`
-        // if !any_filter.is_match(&current_file) {
-        //     continue;
-        // }
-
         let mut ast = syn::parse_file(&current_file).unwrap();
 
         // Get all struct we might work with here
@@ -44,49 +34,14 @@ pub fn load_and_patch_any(out_dir: &Path) -> BTreeMap<String, (File, BTreeMap<St
                 // Find any fields and replace with generics
                 let fields = fields_as_named(&mut item.fields).unwrap();
                 for field in fields.named.iter_mut() {
-                    if let Some((ty, path)) = is_important(field) {
+                    if let Some((_, path)) = is_important(field) {
                         if path.path.segments.last().unwrap().ident == "Any" {
                             // Set struct generics
                             let generic = GENERICS[item.generics.params.len()];
-                            path.path = gen_generic(generic);
+                            path.path = gen_any(generic);
                             item.generics
                                 .params
                                 .push(GenericParam::Type(gen_type_param(generic)));
-
-                            // Set serialization function
-                            let serde_path = match ty {
-                                FoundEnclosure::Option => "option",
-                                FoundEnclosure::Vec => "vec",
-                            };
-
-                            let mut token_stream = TokenStream::new();
-                            token_stream.append(Ident::new("serialize_with", Span::call_site()));
-                            token_stream.append(Punct::new('=', Spacing::Alone));
-                            token_stream.append(Literal::string(&format!(
-                                "crate::any::{}::serialize",
-                                serde_path
-                            )));
-                            token_stream.append(Punct::new(',', Spacing::Alone));
-                            token_stream.append(Ident::new("deserialize_with", Span::call_site()));
-                            token_stream.append(Punct::new('=', Spacing::Alone));
-                            token_stream.append(Literal::string(&format!(
-                                "crate::any::{}::deserialize",
-                                serde_path
-                            )));
-
-                            field.attrs.push(Attribute {
-                                pound_token: Default::default(),
-                                style: AttrStyle::Outer,
-                                bracket_token: Default::default(),
-                                meta: Meta::List(MetaList {
-                                    path: syn::Path {
-                                        leading_colon: None,
-                                        segments: create_punctuated(vec!["serde"]),
-                                    },
-                                    delimiter: MacroDelimiter::Paren(Paren::default()),
-                                    tokens: token_stream,
-                                }),
-                            });
                         }
                     }
                 }
